+++ webui/src/components/primitives/primitives.test.tsx
@@ -1,7 +1,7 @@
 import { render, screen } from '@testing-library/react';
 import { describe, expect, it } from 'vitest';
 
-import { Show } from './show';
+import { Notice, Show } from './primitives';
 
 describe('Show', () => {
   it('renders children when the condition is true', () => {
@@ -29,5 +29,24 @@ describe('Show', () => {
     render(<Show when="Ada">{(name) => <span>{name}</span>}</Show>);
 
     expect(screen.getByText('Ada')).toBeInTheDocument();
+  });
+});
+
+describe('Notice', () => {
+  it('renders as a note by default', () => {
+    render(<Notice>Fallback message</Notice>);
+
+    expect(screen.getByText('Fallback message')).toHaveAttribute('role', 'note');
+    expect(screen.getByText('Fallback message')).toHaveAttribute('data-tone', 'info');
+  });
+
+  it('supports tone overrides', () => {
+    render(
+      <Notice tone="warning">
+        <span>Provider fallback</span>
+      </Notice>,
+    );
+
+    expect(screen.getByRole('note')).toHaveAttribute('data-tone', 'warning');
   });
 });
